feat: Implement product filtering, sorting, and grid display
- Added ProductFilters component for searching, filtering by categories and brands, and setting price ranges.
- Introduced ProductSortBar for sorting products based on various criteria.
- Created ProductGrid to display products in a responsive grid layout with load more functionality.
- Developed ProductHeader for consistent product page header.
- Refactored ProductPage to utilize hooks for product filtering and sorting logic.
- Removed unused components and optimized imports.
- Added slug field to Cosmetic type for better routing.
- Implemented fetchCosmeticBySlug action in cosmeticSlice for fetching products by slug.
- Updated next.config.ts for experimental package imports optimization. #16
