+++ app/(shop)/product/components/ProductHeader.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+const ProductHeader = () => {
+  return (
+    <div className="space-y-4 mb-8">
+      <h1 className="font-inter text-foreground text-3xl font-bold">
+        Danh Mục Sản Phẩm
+      </h1>
+      <p className="text-muted-foreground font-inter text-lg">
+        Tìm kiếm sản phẩm phù hợp với nhu cầu làm đẹp của bạn.
+      </p>
+    </div>
+  );
+};
+
+export default ProductHeader;
