--- components/HoleBackground.tsx
@@ -1,9 +1,0 @@
-import { HoleBackground } from '@/components/animate-ui/components/backgrounds/hole';
-
-const HoleBackgroundComponent = () => {
-  return (
-    <HoleBackground className="absolute inset-0 flex items-center justify-center rounded-xl" />
-  );
-};
-
-export default HoleBackgroundComponent;
