+++ components/address/AddressSelect.tsx
@@ -41,6 +41,7 @@ import {
 } from "@/lib/redux/address/addressSlice";
 import { AppDispatch } from "@/lib/redux/store";
 import { Address, CreateAddressData } from "@/lib/api/address";
+import { toast } from "sonner";
 
 interface AddressSelectProps {
   onAddressChange?: (address: Address | null, index: number) => void;
